+++ src/LinkInstanceFactory.ts
@@ -1,6 +1,5 @@
-import { LinkModel } from "./Common";
-import * as _ from "lodash";
 import { AbstractInstanceFactory } from "./AbstractInstanceFactory";
+import {LinkModel} from "./models/LinkModel";
 /**
  * @author Dylan Vorster
  */
