+++ src/WidgetFactories.ts
@@ -1,5 +1,6 @@
-import { NodeModel, LinkModel } from "./Common";
 import { DiagramEngine } from "./DiagramEngine";
+import {NodeModel} from "./models/NodeModel";
+import {LinkModel} from "./models/LinkModel";
 /**
  * @author Dylan Vorster
  */
