+++ src/defaults/DefaultPortModel.ts
@@ -1,6 +1,6 @@
-import { PortModel } from "../Common";
 import * as _ from "lodash";
 import { AbstractInstanceFactory } from "../AbstractInstanceFactory";
+import {PortModel} from "../models/PortModel";
 
 export class DefaultPortInstanceFactory extends AbstractInstanceFactory<DefaultPortModel> {
 	constructor() {
