+++ src/widgets/LinkLayerWidget.tsx
@@ -1,9 +1,8 @@
 import * as React from "react";
-import { DiagramModel } from "../DiagramModel";
 import { DiagramEngine } from "../DiagramEngine";
-import { PointModel } from "../Common";
 import { LinkWidget } from "./LinkWidget";
 import * as _ from "lodash";
+import {PointModel} from "../models/PointModel";
 
 export interface LinkLayerProps {
 	diagramEngine: DiagramEngine;
